refactor; emptying template utils
